Update README.md
The Spotify Music Recommender is a web application built with Streamlit that utilizes the Spotify API to generate music recommendations based on a seed track. Users can input a Spotify track ID, and the application fetches information about the track and provides a list of recommended tracks with their respective artists. The recommendations are obtained using the Spotify API's recommendation feature, offering users a quick and easy way to discover new music based on their favorite tracks.